mcp-servers/mdcalc-automation-mcp/requirements.txt
# MDCalc MCP Server Dependencies
# Remote MCP server with OAuth 2.1 authentication

# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# OAuth token validation
python-jose[cryptography]==3.3.0
httpx==0.27.0

# Configuration management
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0

# Browser automation (existing)
playwright==1.40.0

# Data processing (existing)
pandas>=2.1.4
numpy>=1.24.3

# Web scraping (existing)
beautifulsoup4>=4.12.0
lxml>=4.9.3

# JSON & validation (existing)
jsonschema>=4.20.0

# Async utilities
aiohttp>=3.9.0
